+++ tests/test_dialog.py
@@ -13,8 +13,8 @@ class TestDialog(BaseTestCase):
     async def test_alert(self):
         def dialog_test(dialog):
             self.assertEqual(dialog.type, 'alert')
-            self.assertEqual(dialog.defaultValue(), '')
-            self.assertEqual(dialog.message(), 'yo')
+            self.assertEqual(dialog.defaultValue, '')
+            self.assertEqual(dialog.message, 'yo')
             asyncio.ensure_future(dialog.accept())
         self.page.on('dialog', dialog_test)
         await self.page.evaluate('() => alert("yo")')
@@ -23,8 +23,8 @@ class TestDialog(BaseTestCase):
     async def test_prompt(self):
         def dialog_test(dialog):
             self.assertEqual(dialog.type, 'prompt')
-            self.assertEqual(dialog.defaultValue(), 'yes.')
-            self.assertEqual(dialog.message(), 'question?')
+            self.assertEqual(dialog.defaultValue, 'yes.')
+            self.assertEqual(dialog.message, 'question?')
             asyncio.ensure_future(dialog.accept('answer!'))
         self.page.on('dialog', dialog_test)
         answer = await self.page.evaluate('() => prompt("question?", "yes.")')